+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## [0.8.3] - 2019-01-16
+### Added
+- Admin Panel: New Gorgeous Light Theme for Admin panel!
+- Admin Panel: Settings Manager - ability to change admin panel theme - added.
+- Admin Panel: Settings Manager - Select dropdown for cache driver - added.
+- Flextype Core: Cache - new cache driver Array - added.
+- Flextype Core: Cache - new cache driver SQLite3 - added.
+- Flextype Core: Cache - new cache driver Zend - added.
+
+### Changed
+- Flextype Core: Symfony Yaml updated to 4.2.2.
+- Admin Panel: Settings Manager - cache settings editing - improvements.
+- Flextype Core: default settings - refactoring.
+
+### Fixed
+- Flextype Core: Cache - issue with selecting current driver_name - fixed.
+- Admin Panel: Dark Theme fixes.
+- Admin Panel: Translates fixes.
+- Admin Panel: Typo fixes.
+
 ## [0.8.2] - 2019-01-09
 ### Added
 - Admin Panel: ClipboardJS added!